+++ modules/justin_version.py
@@ -1,3 +1,3 @@
 # This file must by valid Python AND valid Bash!
-justinVersion='01.06.02'
+justinVersion='01.06.03'